Move device update scripts to an independent file
