+++ pyproject.toml
@@ -54,4 +54,4 @@ build-backend = "poetry.core.masonry.api"
 
 [tool.poetry.scripts]
 stream_image_capture = "miniscope_io.stream_daq:main"
-updateDevice = "miniscope_io.stream_daq:updateDevice"
+updateDevice = "miniscope_io.device_update:updateDevice"
